Added a target
Now there is a small green target for the orb to collect. When it
passes over the target the orb turns a yellow color.

I’m imagining each target collected giving the player a point, and each
return to the original position subtracting a point.

On a side note, the game is starting to feel like it could be fun to
play. It’s not that easy to collect the target (even at the low speed
it currently has) without accidentally sending yourself back to the
spawn point.