Made score counter work
You may want to look at the code to see if I did this in a really poor
way. But I gave each object a property “exists” and as soon as the orb
touches it, it disappears and the score either goes up or down
depending on whether or not its an obstacle or a target.