Added a way to beat the level
When you collect all the targets, the level freezes and displays the
user score. Right now the final score is a function of time and points
(+1 for each target, -1 for each obstacle). This means that for two of
each the minimum score is 0 and the maximum score is limited by the
game time (final_score = 100*points/time, this is an equation we can
fiddle with)