Level Four
This is probably the hardest level yet. I also added a new parameter
for the obstacles/targets that lets you specify their starting angle.
That will make the other levels broken until I pop that on in but it
gives more freedom in level design.

At this point it’s probably also worth it to combine the two objects
(target and obstacle) because they more or less function exactly the
same at this point.